Move Table<Node> impl from table.rs to node_utils.rs
table.rs defines a generic 2D grid Table<T> but previously imported
sxd_xpath::nodeset::Node and held an impl block specific to that type.
This created a layer inversion: the generic data structure depended on
the XPath parser library.

Move the to_string_table and to_string_table_with_header methods into
node_utils.rs, which already owns the sxd_xpath dependency. table.rs
now contains only generic T-agnostic code with no sxd_xpath import.

No behavior change; all existing tests pass.
